+++ src/scripts/ui_builder.js
@@ -383,13 +383,13 @@ var UI_BUILDER = (($)=>{
                 evt.preventDefault()
                 if(MODE === MODE_SETTINGS){
                     this.openSettings(evt)
-                } else if(MODE === MODE_MOVE && (evt.originalEvent.button === 0 || evt.originalEvent instanceof TouchEvent)){
+                } else if(MODE === MODE_MOVE && (evt.originalEvent.button === 0 || (UI.supportsTouch() && evt.originalEvent instanceof TouchEvent)) ){
                     this.activateDrag(evt)
-                } else if (MODE === MODE_RESIZE && (evt.originalEvent.button === 0 || evt.originalEvent instanceof TouchEvent)){
+                } else if (MODE === MODE_RESIZE && (evt.originalEvent.button === 0 || (UI.supportsTouch() && evt.originalEvent instanceof TouchEvent)) ){
                     this.activateResize(evt)
-                } else if (MODE === MODE_DELETE && (evt.originalEvent.button === 0 || evt.originalEvent instanceof TouchEvent)){
+                } else if (MODE === MODE_DELETE && (evt.originalEvent.button === 0 || (UI.supportsTouch() && evt.originalEvent instanceof TouchEvent)) ){
                     this.delete()
-                } else if (MODE === MODE_ZINDEX && (evt.originalEvent.button === 0 || evt.originalEvent instanceof TouchEvent)){
+                } else if (MODE === MODE_ZINDEX && (evt.originalEvent.button === 0 || (UI.supportsTouch() && evt.originalEvent instanceof TouchEvent)) ){
                     moveElementZindexToFront(this)
                 }
             })
@@ -405,15 +405,28 @@ var UI_BUILDER = (($)=>{
         activateDrag(evt){
             this.deactivate()
 
-            this.offX = window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX) - uiZoom(this.x)
-            this.offY = window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY) - uiZoom(this.y)
+            if(UI.supportsTouch()){
+                this.offX = window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX) - uiZoom(this.x)
+                this.offY = window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY) - uiZoom(this.y)
+            } else {
+                this.offX = window.scrollX + evt.clientX - uiZoom(this.x)
+                this.offY = window.scrollY + evt.clientY - uiZoom(this.y)
+            }
 
-            $(gcontainer).on('mousemove touchmove', (evt)=>{
+            let func = (evt)=>{
                 evt.preventDefault()
-                this.x = uiUnzoom((window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX)) - this.offX)
-                this.y = uiUnzoom((window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY)) - this.offY)
+                if(UI.supportsTouch()){
+                    this.x = uiUnzoom((window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX)) - this.offX)
+                    this.y = uiUnzoom((window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY)) - this.offY)
+                } else {
+                    this.x = uiUnzoom((window.scrollX + evt.clientX) - this.offX)
+                    this.y = uiUnzoom((window.scrollY + evt.clientY) - this.offY)
+                }
                 this.refreshPosition()
-            })
+            }
+
+            $(gcontainer).on('mousemove touchmove', func)
+            this.dom.on('mousemove touchmove', func)
 
             $(gcontainer).on('mouseup touchend touchcancel', ()=>{
                 this.deactivate()
@@ -423,15 +436,28 @@ var UI_BUILDER = (($)=>{
         activateResize(evt){
             this.deactivate()
 
-            this.offX = (window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX)) - uiZoom(this.width)
-            this.offY = (window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY)) - uiZoom(this.height)
+            if(UI.supportsTouch()){
+                this.offX = (window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX)) - uiZoom(this.width)
+                this.offY = (window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY)) - uiZoom(this.height)
+            } else {
+                this.offX = (window.scrollX + evt.clientX) - uiZoom(this.width)
+                this.offY = (window.scrollY + evt.clientY) - uiZoom(this.height)
+            }
 
-            $(gcontainer).on('mousemove touchmove', (evt)=>{
+            let func = (evt)=>{
                 evt.preventDefault()
-                this.width = uiUnzoom((window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX)) - this.offX)
-                this.height = uiUnzoom((window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY)) - this.offY)
+                if(UI.supportsTouch()){
+                    this.width = uiUnzoom((window.scrollX + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientX : evt.clientX)) - this.offX)
+                    this.height = uiUnzoom((window.scrollY + (evt.originalEvent instanceof TouchEvent ? evt.originalEvent.touches[0].clientY : evt.clientY)) - this.offY)
+                } else {
+                    this.width = uiUnzoom((window.scrollX + evt.clientX) - this.offX)
+                    this.height = uiUnzoom((window.scrollY + evt.clientY) - this.offY)
+                }
                 this.refreshPosition()
-            })
+            }
+
+            $(gcontainer).on('mousemove touchmove', func)
+            this.dom.on('mousemove touchmove', func)
 
             $(gcontainer).on('mouseup touchend touchcancel', ()=>{
                 this.deactivate()
@@ -440,6 +466,7 @@ var UI_BUILDER = (($)=>{
 
         deactivate(){
             $(gcontainer).off('mousemove touchmove mouseup touchend touchcancel')
+            this.dom.off('mousemove touchmove mouseup touchend touchcancel')
             this.closeSettings()
         }
